restructure serialisers, add pandas util
