+++ src/cloudmappings/serialisers/__init__.py
@@ -1,0 +1,5 @@
+from cloudmappings.serialisers.serialisation import CloudMappingSerialisation
+
+__all__ = [
+    "CloudMappingSerialisation",
+]
